finally found a test to ensure we're not spending unnecessary CPU cycles
This commit will finally fix the CPU usage issues we experienced. The
last commit about this added warnings which I was able to trigger with
this new test case.

So now I can turn the warnings into assertions (because I have a test to
ensure that this won't break in the future again).

* gutachter-suite.c: the final CPU usage fix
* tests/test-suite.c: make sure we won't trigger this assertion